feat: format date in footer
